// src/app/models/api.ts
export class ApiConfig { 
    localApiUrl: string;
    movieUrl: string;
    searchUrl: string;
    personUrl: string;
    params: string;     
    midImgPath: string;
    smallImgPath: string;
    bigBackPath: string;
    midBackPath: string;
    smallBackPath: string;   
    favoriteApiUrl: string;
    bookmarkApiUrl: string;
}